Renames get-type to getType.
