+++ lib/deref/sync.js
@@ -1,7 +1,7 @@
 var $error = require("./../types/error");
 var pathSyntax = require("falcor-path-syntax");
 var getBoundValue = require("./../get/getBoundValue");
-var getType = require("./../support/get-type");
+var getType = require("./../support/getType");
 
 module.exports = function derefSync(boundPathArg) {
