+++ lib/request/RequestQueue.js
@@ -2,8 +2,9 @@
 var SetRequest = require("./../request/SetRequest");
 
 var prefix = require("./../internal/prefix");
-var getType = require("./../support/get-type");
-var isObject = require("./../support/is-object");
+var getType = require("./../support/getType");
+var isObject = require("./../support/isObject");
+var arrayClone = require("./../support/array-clone");
 var falcorPathUtils = require("falcor-path-utils");
 
 /* eslint-disable no-labels block-scoped-var */
