+++ lib/set/sync.js
@@ -1,7 +1,7 @@
 var $error = require("./../types/error");
 var pathSyntax = require("falcor-path-syntax");
-var getType = require("./../support/get-type");
 var setJsonValuesAsJsonDense = require("./../set/set-json-values-as-json-dense");
+var getType = require("./../support/getType");
 var isObject = require("./../support/isObject");
 var isPathValue = require("./../support/isPathValue");
