--- lib/support/get-type.js
@@ -1,9 +1,0 @@
-var isObject = require("./../support/is-object");
-
-module.exports = function getType(node, anyType) {
-    var type = isObject(node) && node.$type || void 0;
-    if (anyType && type) {
-        return "branch";
-    }
-    return type;
-};
